rules_pip_lock/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    update_tools
#
cachetools==3.1.1         # via google-auth
certifi==2019.6.16        # via requests
chardet==3.0.4            # via requests
click==7.0                # via pip-tools
dill==0.2.9               # via google-gax
enum34==1.1.6             # via google-cloud-language, grpcio
funcsigs==1.0.2           # via mock
future==0.16.0            # via google-gax
futures==3.3.0            # via google-cloud-core, grpcio
google-auth==1.6.3        # via google-cloud-core, google-gax
google-cloud-core==0.27.1  # via google-cloud-language
google-cloud-language==0.29.0
google-gax==0.15.16       # via google-cloud-language
googleapis-common-protos[grpc]==1.6.0  # via google-cloud-core, google-cloud-language, google-gax
grpcio==1.23.0            # via google-gax, googleapis-common-protos
idna==2.8                 # via requests
mock==3.0.5
pg8000==1.12.5            # via testing.postgresql
pip-tools==4.1.0
pkginfo==1.5.0.1
ply==3.8                  # via google-gax
protobuf==3.9.1           # via google-cloud-core, google-gax, googleapis-common-protos
pyasn1-modules==0.2.6     # via google-auth
pyasn1==0.4.7             # via pyasn1-modules, rsa
pymysql==0.9.3            # via testing.mysqld
requests==2.22.0          # via google-cloud-core, google-gax
rsa==4.0                  # via google-auth
six==1.12.0               # via google-auth, google-cloud-core, grpcio, mock, pg8000, pip-tools, protobuf
testing.common.database==2.0.3  # via testing.mysqld, testing.postgresql
testing.mysqld==1.4.0
testing.postgresql==1.3.0
urllib3==1.25.3           # via requests

# The following packages are considered to be unsafe in a requirements file:
pip==19.2.3
setuptools==41.0.1
